wip: refactor the code structure
