+++ src/core/language/docker.rs
@@ -49,7 +49,7 @@ impl Queryable for Dockerfile {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
+    use crate::core::{
         matcher::MatchedItem,
         query::{MetavariableId, Query},
         tree::Tree,
