+++ src/core/util.rs
