+++ src/lib.rs
@@ -1,16 +1,4 @@
 pub mod cli;
-pub mod code;
-pub mod constraint;
-pub mod language;
-pub mod matcher;
-pub mod node;
-pub mod pattern;
-pub mod query;
+pub mod core;
 pub mod reporter;
-pub mod ruleset;
-pub mod string;
-pub mod target;
 pub mod tests;
-pub mod transform;
-pub mod tree;
-pub mod util;
