+++ src/reporter/json.rs
@@ -1,6 +1,9 @@
 use super::Reporter;
-use crate::code::Code;
-use crate::transform::Transformable;
+use crate::core::matcher::MatchedItem;
+use crate::core::ruleset::Rule;
+use crate::core::target::Target;
+use crate::core::transform::Transformable;
+use crate::core::{code::Code, language::Queryable};
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use similar::TextDiff;
@@ -31,10 +34,10 @@ impl<'a, W: std::io::Write> Reporter<'a> for JSONReporter<'a, W> {
         }
     }
 
-    fn add_entry<T: crate::language::Queryable + 'static>(
+    fn add_entry<T: Queryable + 'static>(
         &mut self,
-        target: &crate::target::Target,
-        items: Vec<(&crate::ruleset::Rule, crate::matcher::MatchedItem)>,
+        target: &Target,
+        items: Vec<(&Rule, MatchedItem)>,
     ) -> Result<()> {
         let target_path = if let Some(ref p) = target.path {
             let p = p.canonicalize()?;
